Finish shovel

+++ I2P_MiniProject2/TowerDefense/PlayScene.hpp
@@ -65,7 +65,7 @@ public:
 	std::vector<std::vector<int>> mapDistance;
 	std::list<std::pair<int, float>> enemyWaveData;
 	std::list<int> keyStrokes;
-	Turret* FindTurretType(float x, float y);
+	Turret* FindTurretType(int x, int y);
 	static Engine::Point GetClientSize();
 	explicit PlayScene() = default;
 	void Initialize() override;
